Register s3 in utils.uri not in storage
resolve_uri depends on s3 being registered with pythons urllib
